added form validation
