+++ src/trpc/routes/user/index.ts
@@ -1,5 +1,4 @@
 import {updateUser} from "./user-update"
-import { getUser } from "./user-getUser"
 
-export {updateUser,getUser};
+export {updateUser};
